requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0

# Vector database
qdrant-client>=1.7.0

# Embedding models
FlagEmbedding>=1.2.10  # For BGE-M3 model (includes sentence-transformers)

# Keyword extraction
keybert>=0.9.0  # For KeyBERT-based keyword extraction
sentence-transformers>=2.2.0  # Required by KeyBERT (multilingual models)

# Document parsing
docling>=1.0.0  # For parsing documents (PDF, DOCX, PPTX, etc.)
PyPDF2>=3.0.0  # PDF parsing (Fallback)
python-docx>=0.8.11  # Word document parsing (Fallback)
pandas>=2.0.0  # Excel parsing (Fallback)
openpyxl>=3.1.0  # Excel support

# Database and system
psutil>=5.9.0
aiohttp>=3.9.0         # For asynchronous web requests
beautifulsoup4>=4.12.0 # For parsing HTML
lxml>=4.0.0,<6.0.0 
trafilatura>=2.0.0     # For extracting main content from web pages

# Web framework
fastapi>=0.104.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn>=0.24.0
python-multipart>=0.0.6  # For file upload (Form, File) in FastAPI

# Authentication and OAuth
authlib>=1.2.0  # For Google OAuth 2.0
python-jose[cryptography]>=3.3.0  # For JWT token handling
httpx>=0.25.0  # For async HTTP requests

# LangChain and AI
langchain>=0.3.0 
langchain-community>=0.0.20
langchain-core>=0.1.0
langgraph>=0.0.40
google-generativeai>=0.3.0

# Data visualization
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
pywin32>=305; sys_platform == 'win32'  # For Windows system information collection
keyring>=24.0.0  # For secure token storage
apscheduler>=3.10.0 # For background task scheduling
websocket-client>=1.9.0  # For WebSocket client (real-time notifications)
websockets>=12.0  # For uvicorn WebSocket support (server-side)

# Report Agent - Web Search
ddgs>=1.0.0  # For DuckDuckGo web search (no API key required) - renamed from duckduckgo-search
duckduckgo-search>=6.0.0  # Fallback for older installations

# Report Agent - PDF Generation
reportlab>=4.0.0  # For PDF generation (primary method, platypus module)
markdown2>=2.4.0  # For Markdown to HTML conversion
xhtml2pdf>=0.2.11  # For HTML to PDF conversion (fallback method)

# PyQt6 Frontend (Phase 0+)
PyQt6>=6.6.0  # Main GUI framework
# PyQt6-Charts>=6.6.0  # Uncomment when dashboard charts are needed
# PyQt6-WebEngine  # Avoid for now - PyInstaller bundle issues